requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
qrcode==8.0
setuptools==75.8.0
Werkzeug==3.1.3
wheel==0.45.1
